Move free functions in main to Core methods
These "free" functions usually took both core and pack as
arguments. Core is not very useful without a pack to give it
symbol and type information.

Now Core contains a Pack (publicly), and all of these utility methods
that were free in main.rs are now methods on the Core struct

I think the code looks nicer with this change.