Backtrace when zephyr is not configured for userspace
Turns out that when zephyr is not configured for userspace, the
mode member of the _thread_arch struct is missing. This caused
errors when attempting to backtrace an application without
userspace configured.

This patch checks for the presence of the mode member, and falls
back to a default value for the PC.