Replace filter separator with a '|'
Now that I have used aerology queries more liberally, I can say
with certainty that '=>' is a bad filter separator. In particular,
This conflicts with a *nix shell's redirect output operator, '>',
meaning that aerology will fail silently, writing the error message
to a file specified as the remainder of the query.

Instead, I'm proposing '|' as it will cause bash to treat the rest
of the query as a sequence of commands and they will _probably_ error
in ways that would lead the user to understand the problem.

In both cases, quoting the query is the best option, but '|'
fails more gracefully.